Refactor TokenEscrow tests
* Extracted behavior to a different file
